+++ actix-http/tests/test_server.rs
@@ -37,7 +37,7 @@ fn test_h1() {
             .h1(|_| future::ok::<_, ()>(Response::Ok().finish()))
     });
 
-    let response = srv.block_on(srv.get().send()).unwrap();
+    let response = srv.block_on(srv.get("/").send()).unwrap();
     assert!(response.status().is_success());
 }
 
@@ -55,7 +55,7 @@ fn test_h1_2() {
             .map(|_| ())
     });
 
-    let response = srv.block_on(srv.get().send()).unwrap();
+    let response = srv.block_on(srv.get("/").send()).unwrap();
     assert!(response.status().is_success());
 }
 
@@ -98,7 +98,7 @@ fn test_h2() -> std::io::Result<()> {
             )
     });
 
-    let response = srv.block_on(srv.sget().send()).unwrap();
+    let response = srv.block_on(srv.sget("/").send()).unwrap();
     assert!(response.status().is_success());
     Ok(())
 }
@@ -121,7 +121,7 @@ fn test_h2_1() -> std::io::Result<()> {
             )
     });
 
-    let response = srv.block_on(srv.sget().send()).unwrap();
+    let response = srv.block_on(srv.sget("/").send()).unwrap();
     assert!(response.status().is_success());
     Ok(())
 }
@@ -145,7 +145,7 @@ fn test_h2_body() -> std::io::Result<()> {
             )
     });
 
-    let response = srv.block_on(srv.sget().send_body(data.clone())).unwrap();
+    let response = srv.block_on(srv.sget("/").send_body(data.clone())).unwrap();
     assert!(response.status().is_success());
 
     let body = srv.load_body(response).unwrap();
@@ -437,7 +437,7 @@ fn test_h1_headers() {
         })
     });
 
-    let response = srv.block_on(srv.get().send()).unwrap();
+    let response = srv.block_on(srv.get("/").send()).unwrap();
     assert!(response.status().is_success());
 
     // read response
@@ -482,7 +482,7 @@ fn test_h2_headers() {
         }).map_err(|_| ()))
     });
 
-    let response = srv.block_on(srv.sget().send()).unwrap();
+    let response = srv.block_on(srv.sget("/").send()).unwrap();
     assert!(response.status().is_success());
 
     // read response
@@ -518,7 +518,7 @@ fn test_h1_body() {
         HttpService::build().h1(|_| future::ok::<_, ()>(Response::Ok().body(STR)))
     });
 
-    let response = srv.block_on(srv.get().send()).unwrap();
+    let response = srv.block_on(srv.get("/").send()).unwrap();
     assert!(response.status().is_success());
 
     // read response
@@ -541,7 +541,7 @@ fn test_h2_body2() {
             )
     });
 
-    let response = srv.block_on(srv.sget().send()).unwrap();
+    let response = srv.block_on(srv.sget("/").send()).unwrap();
     assert!(response.status().is_success());
 
     // read response
@@ -555,7 +555,7 @@ fn test_h1_head_empty() {
         HttpService::build().h1(|_| ok::<_, ()>(Response::Ok().body(STR)))
     });
 
-    let response = srv.block_on(srv.head().send()).unwrap();
+    let response = srv.block_on(srv.head("/").send()).unwrap();
     assert!(response.status().is_success());
 
     {
@@ -586,7 +586,7 @@ fn test_h2_head_empty() {
             )
     });
 
-    let response = srv.block_on(srv.shead().send()).unwrap();
+    let response = srv.block_on(srv.shead("/").send()).unwrap();
     assert!(response.status().is_success());
     assert_eq!(response.version(), http::Version::HTTP_2);
 
@@ -611,7 +611,7 @@ fn test_h1_head_binary() {
         })
     });
 
-    let response = srv.block_on(srv.head().send()).unwrap();
+    let response = srv.block_on(srv.head("/").send()).unwrap();
     assert!(response.status().is_success());
 
     {
@@ -646,7 +646,7 @@ fn test_h2_head_binary() {
             )
     });
 
-    let response = srv.block_on(srv.shead().send()).unwrap();
+    let response = srv.block_on(srv.shead("/").send()).unwrap();
     assert!(response.status().is_success());
 
     {
@@ -668,7 +668,7 @@ fn test_h1_head_binary2() {
         HttpService::build().h1(|_| ok::<_, ()>(Response::Ok().body(STR)))
     });
 
-    let response = srv.block_on(srv.head().send()).unwrap();
+    let response = srv.block_on(srv.head("/").send()).unwrap();
     assert!(response.status().is_success());
 
     {
@@ -695,7 +695,7 @@ fn test_h2_head_binary2() {
             )
     });
 
-    let response = srv.block_on(srv.shead().send()).unwrap();
+    let response = srv.block_on(srv.shead("/").send()).unwrap();
     assert!(response.status().is_success());
 
     {
@@ -719,7 +719,7 @@ fn test_h1_body_length() {
         })
     });
 
-    let response = srv.block_on(srv.get().send()).unwrap();
+    let response = srv.block_on(srv.get("/").send()).unwrap();
     assert!(response.status().is_success());
 
     // read response
@@ -747,7 +747,7 @@ fn test_h2_body_length() {
             )
     });
 
-    let response = srv.block_on(srv.sget().send()).unwrap();
+    let response = srv.block_on(srv.sget("/").send()).unwrap();
     assert!(response.status().is_success());
 
     // read response
@@ -768,7 +768,7 @@ fn test_h1_body_chunked_explicit() {
         })
     });
 
-    let response = srv.block_on(srv.get().send()).unwrap();
+    let response = srv.block_on(srv.get("/").send()).unwrap();
     assert!(response.status().is_success());
     assert_eq!(
         response
@@ -810,7 +810,7 @@ fn test_h2_body_chunked_explicit() {
             )
     });
 
-    let response = srv.block_on(srv.sget().send()).unwrap();
+    let response = srv.block_on(srv.sget("/").send()).unwrap();
     assert!(response.status().is_success());
     assert!(!response.headers().contains_key(header::TRANSFER_ENCODING));
 
@@ -830,7 +830,7 @@ fn test_h1_body_chunked_implicit() {
         })
     });
 
-    let response = srv.block_on(srv.get().send()).unwrap();
+    let response = srv.block_on(srv.get("/").send()).unwrap();
     assert!(response.status().is_success());
     assert_eq!(
         response
@@ -862,7 +862,7 @@ fn test_h1_response_http_error_handling() {
         }))
     });
 
-    let response = srv.block_on(srv.get().send()).unwrap();
+    let response = srv.block_on(srv.get("/").send()).unwrap();
     assert_eq!(response.status(), http::StatusCode::INTERNAL_SERVER_ERROR);
 
     // read response
@@ -895,7 +895,7 @@ fn test_h2_response_http_error_handling() {
             )
     });
 
-    let response = srv.block_on(srv.sget().send()).unwrap();
+    let response = srv.block_on(srv.sget("/").send()).unwrap();
     assert_eq!(response.status(), http::StatusCode::INTERNAL_SERVER_ERROR);
 
     // read response
@@ -910,7 +910,7 @@ fn test_h1_service_error() {
             .h1(|_| Err::<Response, Error>(error::ErrorBadRequest("error")))
     });
 
-    let response = srv.block_on(srv.get().send()).unwrap();
+    let response = srv.block_on(srv.get("/").send()).unwrap();
     assert_eq!(response.status(), http::StatusCode::INTERNAL_SERVER_ERROR);
 
     // read response
@@ -934,7 +934,7 @@ fn test_h2_service_error() {
             )
     });
 
-    let response = srv.block_on(srv.sget().send()).unwrap();
+    let response = srv.block_on(srv.sget("/").send()).unwrap();
     assert_eq!(response.status(), http::StatusCode::INTERNAL_SERVER_ERROR);
 
     // read response